chore: label delete validation

+++ QB_Brain/Areas/IT/Controllers/LabelController.cs
@@ -1,4 +1,5 @@
 ﻿using Base.Entities;
+using Base.Services;
 using Base.Services.Interfaces;
 using IT_Web.Areas.IT.VIewModels.Label;
 using IT_Web.Extensions;
@@ -12,11 +13,13 @@ public class LabelController : Controller
 {
     private readonly ILogger<LabelController> _logger;
     private readonly ILabelService _labelService;
+    private IIssueLabelService _issueLabelService;
 
-    public LabelController(ILogger<LabelController> logger, ILabelService labelService)
+    public LabelController(ILogger<LabelController> logger, ILabelService labelService, IIssueLabelService issueLabelService)
     {
         _logger = logger;
         _labelService = labelService;
+        _issueLabelService = issueLabelService;
     }
 
     [HttpGet]
@@ -84,6 +87,8 @@ public class LabelController : Controller
     {
         try
         {
+            var issueLabel = await _issueLabelService.CheckIfLabelInUse(id);
+            if (issueLabel) throw new Exception("Can't to delete label. Label is assigned to an issue");
             await _labelService.DeleteLabel(id);
             return RedirectToAction(nameof(New));
         }
